--- notion_extensions/base/props/block.py
@@ -1,2609 +1,0 @@
-import sys
-from typing import Dict, List, Optional, Union
-
-from .common import BaseProps, Emoji, FileObject, Icon, Text, RichText
-from ..utils import parse_id
-
-if sys.version_info >= (3, 8):  # "from typing" in Python 3.9 and earlier
-    from typing import Literal
-else:
-    from typing_extensions import Literal
-if sys.version_info > (3, 9):  # "from typing_extensions" in Python 3.9 and earlier
-    from typing import TypeAlias
-else:
-    from typing_extensions import TypeAlias
-
-__all__ = [
-    "Block",
-    "Children",
-    "Paragraph",
-    "Heading1",
-    "Heading2",
-    "Heading3",
-    "Callout",
-    "Quote",
-    "BulletedListItem",
-    "BulletedList",
-    "NumberedListItem",
-    "NumberedList",
-    "ToDo",
-    "ToDoList",
-    "Toggle",
-    "Code",
-    "Embed",
-    "Image",
-    "Video",
-]
-
-BLOCK_TYPES = Literal[
-    "paragraph",
-    "heading_1",
-    "heading_2",
-    "heading_3",
-    "bulleted_list_item",
-    "numbered_list_item",
-    "to_do",
-    "toggle",
-    "child_page",
-    "child_database",
-    "embed",
-    "image",
-    "video",
-    "file",
-    "pdf",
-    "bookmark",
-    "callout",
-    "quote",
-    "equation",
-    "divider",
-    "table_of_contents",
-    "column",
-    "column_list",
-    "link_preview",
-    "synced_block",
-    "template",
-    "link_to_page",
-    "table",
-    "table_row",
-    "unsupported",
-    "children",
-]
-
-LANGUAGES: TypeAlias = Literal[
-    "abap",
-    "arduino",
-    "bash",
-    "basic",
-    "c",
-    "clojure",
-    "coffeescript",
-    "c++",
-    "c#",
-    "css",
-    "dart",
-    "diff",
-    "docker",
-    "elixir",
-    "elm",
-    "erlang",
-    "flow",
-    "fortran",
-    "f#",
-    "gherkin",
-    "glsl",
-    "go",
-    "graphql",
-    "groovy",
-    "haskell",
-    "html",
-    "java",
-    "javascript",
-    "json",
-    "julia",
-    "kotlin",
-    "latex",
-    "less",
-    "lisp",
-    "livescript",
-    "lua",
-    "makefile",
-    "markdown",
-    "markup",
-    "matlab",
-    "mermaid",
-    "nix",
-    "objective-c",
-    "ocaml",
-    "pascal",
-    "perl",
-    "php",
-    "plain text",
-    "powershell",
-    "prolog",
-    "protobuf",
-    "python",
-    "r",
-    "reason",
-    "ruby",
-    "rust",
-    "sass",
-    "scala",
-    "scheme",
-    "scss",
-    "shell",
-    "sql",
-    "swift",
-    "typescript",
-    "vb.net",
-    "verilog",
-    "vhdl",
-    "visual basic",
-    "webassembly",
-    "xml",
-    "yaml",
-    "java/c/c++/c#",
-]
-
-EMBED_PLATFORMS = [
-    "Framer",
-    "Twitter (tweets)",
-    "Google Drive documents",
-    "Gist",
-    "Figma",
-    "Invision,",
-    "Loom",
-    "Typeform",
-    "Codepen",
-    "PDFs",
-    "Google Maps",
-    "Whimisical",
-    "Miro",
-    "Abstract",
-    "excalidraw",
-    "Sketch",
-    "Replit",
-]
-
-IMAGE_EXT: List[str] = [
-    "png",
-    "jpg",
-    "jpeg",
-    "gif",
-    "tif",
-    "tiff",
-    "bmp",
-    "svg",
-    "heic",
-]
-
-
-class Block(BaseProps):
-    def __init__(self):
-        super().__init__()
-
-
-class Children(Block):
-    """
-    Children
-    Children property values of block
-
-    Attributes
-    ----------
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, List[Block]] = {"children": []}
-
-    def __init__(
-        self,
-        *block: Block,
-    ):
-        super().__init__()
-        self.__blocks = list(block)
-        self["children"] = self.__blocks
-
-    def __add__(self, other: Union[Block, List[Block]]):
-        if isinstance(other, list):
-            self.extend(other)
-            return self
-        self.append(other)
-        return self
-
-    def __iadd__(self, other: Union[Block, List[Block]]):
-        return self.__add__(other)
-
-    def append(self, block: Block) -> None:
-        """
-        append(block: Block)
-            Append Block to existing list of Block
-
-        Parameters
-        ----------
-        block : Block
-            Block you append to Children
-        """
-        self.__blocks.append(block)
-        self["children"] = self.__blocks
-
-    def extend(self, blocks: List[Block]) -> None:
-        """
-        extens(blocks: List[Block])
-            Append Block to existing list of Block
-
-        Parameters
-        ----------
-        blocks : list of Block
-            List of block you append to Children
-        """
-        self.__blocks.extend(blocks)
-        self["children"] = self.__blocks
-
-    def insert(self, index: int, block: Block) -> None:
-        """
-        insert(index: int, block: Block)
-            Append block to existing list of Block
-
-        Parameters
-        ----------
-        index : int
-            Index you insert Block into Children
-        block : Block
-            Block you insert into Children
-        """
-        self.__blocks.insert(index, block)
-        self["children"] = self.__blocks
-
-    def pop(self, index=None):
-        """
-        pop(Block: Block)
-            Pop Block to existing list of Block
-
-        Parameters
-        ----------
-        index : int, default=None
-            Block you pop from RichBlock
-        """
-        item = self.__blocks.pop(index)
-        self["children"] = self.__blocks
-        return item
-
-
-class Paragraph(Block):
-    """
-    Paragraph
-    Paragraph property values of block
-
-    Attributes
-    ----------
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "object": "block",
-        "type": "paragraph",
-        "paragraph": {
-            "text": [],
-        },
-    }
-
-    def __init__(
-        self,
-        *text: Union[Text, RichText],
-    ):
-        super().__init__()
-        base = []
-        for t in text:
-            if isinstance(t, RichText):
-                base.extend(list(t[t.key]))
-            elif isinstance(t, Text):
-                base.append(t)
-            else:
-                raise ValueError(
-                    f"Expected type is `RichText` or `Text`, but {type(t)} is given"
-                )
-        self.__texts = RichText(key="text", *base)
-
-        self.update(
-            {
-                "object": "block",
-                "type": "paragraph",
-                "paragraph": self.__texts,
-            },
-        )
-
-    def __add__(self, other: Union[Text, List[Text]]):
-        if isinstance(other, list):
-            self.extend(other)
-            return self
-        self.append(other)
-        return self
-
-    def __iadd__(self, other: Union[Text, List[Text]]):
-        return self.__add__(other)
-
-    def append(self, text: Text) -> None:
-        """
-        append(text: Text)
-            Append Text to existing list of Text
-
-        Parameters
-        ----------
-        text : Text
-            Text you append to RichText
-        """
-        self.__texts.append(text)
-        self["paragraph"] = self.__texts
-
-    def extend(self, texts: List[Text]) -> None:
-        """
-        extens(texts: Text)
-            Append Text to existing list of Text
-
-        Parameters
-        ----------
-        text : list of Text
-            List of text you append to RichText
-        """
-        self.__texts.extend(texts)
-        self["paragraph"] = self.__texts
-
-    def insert(self, index: int, text: Text) -> None:
-        """
-        insert(index: int, text: Text)
-            Append Text to existing list of Text
-
-        Parameters
-        ----------
-        index : int
-            Index you insert Text into
-        text : Text
-            Text you insert into RichText
-        """
-        self.__texts.insert(index, text)
-        self["paragraph"] = self.__texts
-
-    def pop(self, index=None):
-        """
-        pop(text: Text)
-            Pop Text to existing list of Text
-
-        Parameters
-        ----------
-        index : int, default=None
-            Text you pop from RichText
-        """
-        item = self.__texts.pop(index)
-        self["paragraph"] = self.__texts
-        return item
-
-
-class Heading1(Block):
-    """
-    Heading1
-    Heading1 property values of block
-
-    Attributes
-    ----------
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "object": "block",
-        "type": "heading_1",
-        "heading_1": {
-            "text": [],
-        },
-    }
-
-    def __init__(
-        self,
-        *text: Union[RichText, Text],
-    ):
-        super().__init__()
-        base = []
-        for t in text:
-            if isinstance(t, RichText):
-                base.extend(list(t[t.key]))
-            elif isinstance(t, Text):
-                base.append(t)
-            else:
-                raise ValueError(
-                    f"Expected type is `RichText` or `Text`, but {type(t)} is given"
-                )
-        self.__texts = RichText(key="text", *base)
-
-        self.update(
-            {
-                "object": "block",
-                "type": "heading_1",
-                "heading_1": self.__texts,
-            },
-        )
-
-    def __add__(self, other: Union[Text, List[Text]]):
-        if isinstance(other, list):
-            self.extend(other)
-            return self
-        self.append(other)
-        return self
-
-    def __iadd__(self, other: Union[Text, List[Text]]):
-        return self.__add__(other)
-
-    def append(self, text: Text) -> None:
-        """
-        append(text: Text)
-            Append Text to existing list of Text
-
-        Parameters
-        ----------
-        text : Text
-            Text you append to RichText
-        """
-        self.__texts.append(text)
-        self["heading_1"] = self.__texts
-
-    def extend(self, texts: List[Text]) -> None:
-        """
-        extens(texts: Text)
-            Append Text to existing list of Text
-
-        Parameters
-        ----------
-        text : list of Text
-            List of text you append to RichText
-        """
-        self.__texts.extend(texts)
-        self["heading_1"] = self.__texts
-
-    def insert(self, index: int, text: Text) -> None:
-        """
-        insert(index: int, text: Text)
-            Append Text to existing list of Text
-
-        Parameters
-        ----------
-        index : int
-            Index you insert Text into
-        text : Text
-            Text you insert into RichText
-        """
-        self.__texts.insert(index, text)
-        self["heading_1"] = self.__texts
-
-    def pop(self, index=None):
-        """
-        pop(text: Text)
-            Pop Text to existing list of Text
-
-        Parameters
-        ----------
-        index : int, default=None
-            Text you pop from RichText
-        """
-        item = self.__texts.pop(index)
-        self["heading_1"] = self.__texts
-        return item
-
-
-class Heading2(Block):
-    """
-    Heading2
-    Heading2 property values of block
-
-    Attributes
-    ----------
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "object": "block",
-        "type": "heading_2",
-        "heading_2": {
-            "text": [],
-        },
-    }
-
-    def __init__(
-        self,
-        *text: Union[RichText, Text],
-    ):
-        super().__init__()
-        base = []
-        for t in text:
-            if isinstance(t, RichText):
-                base.extend(list(t[t.key]))
-            elif isinstance(t, Text):
-                base.append(t)
-            else:
-                raise ValueError(
-                    f"Expected type is `RichText` or `Text`, but {type(t)} is given"
-                )
-        self.__texts = RichText(key="text", *base)
-
-        self.update(
-            {
-                "object": "block",
-                "type": "heading_2",
-                "heading_2": self.__texts,
-            },
-        )
-
-    def __add__(self, other: Union[Text, List[Text]]):
-        if isinstance(other, list):
-            self.extend(other)
-            return self
-        self.append(other)
-        return self
-
-    def __iadd__(self, other: Union[Text, List[Text]]):
-        return self.__add__(other)
-
-    def append(self, text: Text) -> None:
-        """
-        append(text: Text)
-            Append Text to existing list of Text
-
-        Parameters
-        ----------
-        text : Text
-            Text you append to RichText
-        """
-        self.__texts.append(text)
-        self["heading_2"] = self.__texts
-
-    def extend(self, texts: List[Text]) -> None:
-        """
-        extens(texts: Text)
-            Append Text to existing list of Text
-
-        Parameters
-        ----------
-        text : list of Text
-            List of text you append to RichText
-        """
-        self.__texts.extend(texts)
-        self["heading_2"] = self.__texts
-
-    def insert(self, index: int, text: Text) -> None:
-        """
-        insert(index: int, text: Text)
-            Append Text to existing list of Text
-
-        Parameters
-        ----------
-        index : int
-            Index you insert Text into
-        text : Text
-            Text you insert into RichText
-        """
-        self.__texts.insert(index, text)
-        self["heading_2"] = self.__texts
-
-    def pop(self, index=None):
-        """
-        pop(text: Text)
-            Pop Text to existing list of Text
-
-        Parameters
-        ----------
-        index : int, default=None
-            Text you pop from RichText
-        """
-        item = self.__texts.pop(index)
-        self["heading_2"] = self.__texts
-        return item
-
-
-class Heading3(Block):
-    """
-    Heading3
-    Heading3 property values of block
-
-    Attributes
-    ----------
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "object": "block",
-        "type": "heading_3",
-        "heading_3": {
-            "text": [],
-        },
-    }
-
-    def __init__(
-        self,
-        *text: Union[RichText, Text],
-    ):
-        super().__init__()
-        base = []
-        for t in text:
-            if isinstance(t, RichText):
-                base.extend(list(t[t.key]))
-            elif isinstance(t, Text):
-                base.append(t)
-            else:
-                raise ValueError(
-                    f"Expected type is `RichText` or `Text`, but {type(t)} is given"
-                )
-        self.__texts = RichText(key="text", *base)
-
-        self.update(
-            {
-                "object": "block",
-                "type": "heading_3",
-                "heading_3": self.__texts,
-            },
-        )
-
-    def __add__(self, other: Union[Text, List[Text]]):
-        if isinstance(other, list):
-            self.extend(other)
-            return self
-        self.append(other)
-        return self
-
-    def __iadd__(self, other: Union[Text, List[Text]]):
-        return self.__add__(other)
-
-    def append(self, text: Text) -> None:
-        """
-        append(text: Text)
-            Append Text to existing list of Text
-
-        Parameters
-        ----------
-        text : Text
-            Text you append to RichText
-        """
-        self.__texts.append(text)
-        self["heading_3"] = self.__texts
-
-    def extend(self, texts: List[Text]) -> None:
-        """
-        extens(texts: Text)
-            Append Text to existing list of Text
-
-        Parameters
-        ----------
-        text : list of Text
-            List of text you append to RichText
-        """
-        self.__texts.extend(texts)
-        self["heading_3"] = self.__texts
-
-    def insert(self, index: int, text: Text) -> None:
-        """
-        insert(index: int, text: Text)
-            Append Text to existing list of Text
-
-        Parameters
-        ----------
-        index : int
-            Index you insert Text into
-        text : Text
-            Text you insert into RichText
-        """
-        self.__texts.insert(index, text)
-        self["heading_3"] = self.__texts
-
-    def pop(self, index=None):
-        """
-        pop(text: Text)
-            Pop Text to existing list of Text
-
-        Parameters
-        ----------
-        index : int, default=None
-            Text you pop from RichText
-        """
-        item = self.__texts.pop(index)
-        self["heading_3"] = self.__texts
-        return item
-
-
-class Callout(Block):
-    """
-    Callout
-    Callout property values of block
-
-    Attributes
-    ----------
-    text : RichText
-        text
-    icon : Icon
-        icon
-    children : Children
-        children
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "type": "callout",
-        "callout": {
-            "text": [],
-        },
-    }
-
-    def __init__(
-        self,
-        *text: Union[Text, RichText],
-        icon: Icon = Icon(Emoji("")),
-        children: Optional[Children] = None,
-    ):
-        """
-        Parameters
-        ----------
-        *text : Text or RichText
-            text
-        icon : Icon, defalut=Icon(Emoji(""))
-            icon
-        children : Children, optional
-            children
-        """
-        super().__init__()
-        self["callout"].update(icon)  # Add Icon
-        base = []  # Aggregate Texts
-        for t in text:
-            if isinstance(t, RichText):
-                base.extend(list(t[t.key]))
-            elif isinstance(t, Text):
-                base.append(t)
-            else:
-                raise ValueError(
-                    f"Expected type is `RichText` or `Text`, but {type(t)} is given"
-                )
-        self.__text = RichText(key="text", *base)
-        self["callout"].update(self.__text)  # Add Texts with RichText Style
-        if children is not None:
-            self["callout"].update(children)  # if children exists, Add Chilren
-
-    @property
-    def text(self) -> RichText:
-        return self.__text
-
-    @text.setter
-    def text(self, value: RichText) -> None:
-        if value.key != "text":
-            raise ValueError("RichText's key is must be `text`")
-        self.__text = value
-        self["callout"].update(self.__text)
-
-    @property
-    def icon(self) -> Icon:
-        return self["callout"]["icon"]
-
-    @icon.setter
-    def icon(self, value: Icon) -> None:
-        self["callout"].update(value)
-
-    @property
-    def children(self) -> Children:
-        return self["callout"]["children"]
-
-    @children.setter
-    def children(self, value: Children) -> None:
-        self["callout"].update(value)
-
-
-class Quote(Block):
-    """
-    Quote
-    Quote property values of block
-
-    Attributes
-    ----------
-    text : RichText
-        text
-    children : Children
-        children
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "type": "quote",
-        "quote": {
-            "text": [],
-        },
-    }
-
-    def __init__(
-        self,
-        *text: Union[Text, RichText],
-        children: Optional[Children] = None,
-    ):
-        """
-        Parameters
-        ----------
-        *text : Text or RichText
-            text
-        children : Children, optional
-            children
-        """
-        super().__init__()
-        base = []  # Aggregate Texts
-        for t in text:
-            if isinstance(t, RichText):
-                base.extend(list(t[t.key]))
-            elif isinstance(t, Text):
-                base.append(t)
-            else:
-                raise ValueError(
-                    f"Expected type is `RichText` or `Text`, but {type(t)} is given"
-                )
-        self.__text = RichText(key="text", *base)
-        self["quote"].update(self.__text)  # Add Texts with RichText Style
-        if children is not None:
-            self["quote"].update(children)  # if children exists, Add Chilren
-
-    @property
-    def text(self) -> RichText:
-        return self.__text
-
-    @text.setter
-    def text(self, value: RichText) -> None:
-        if value.key != "text":
-            raise ValueError("RichText's key is must be `text`")
-        self.__text = value
-        self["quote"].update(self.__text)
-
-    @property
-    def children(self) -> Children:
-        return self["quote"]["children"]
-
-    @children.setter
-    def children(self, value: Children) -> None:
-        self["quote"].update(value)
-
-
-class BulletedListItem(Block):
-    """
-    BulletedListItem
-    BulletedListItem property values of block
-
-    Attributes
-    ----------
-    *text : Text or RichText
-        text
-    children : Children, optional
-        children
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "type": "bulleted_list_item",
-        "bulleted_list_item": {
-            "text": [],
-        },
-    }
-
-    def __init__(
-        self,
-        *text: Union[Text, RichText],
-        children: Optional[Children] = None,
-    ):
-        """
-        Parameters
-        ----------
-        *text : Text or RichText
-            text
-        children : Children, optional
-            children
-        """
-        super().__init__()
-        base = []  # Aggregate Texts
-        for t in text:
-            if isinstance(t, RichText):
-                base.extend(list(t[t.key]))
-            elif isinstance(t, Text):
-                base.append(t)
-            else:
-                raise ValueError(
-                    f"Expected type is `RichText` or `Text`, but {type(t)} is given"
-                )
-        self.__text = RichText(key="text", *base)
-        self["bulleted_list_item"].update(self.__text)  # Add Texts with RichText Style
-        if children is not None:
-            self["bulleted_list_item"].update(
-                children
-            )  # if children exists, Add Chilren
-
-    @property
-    def text(self) -> RichText:
-        return self.__text
-
-    @text.setter
-    def text(self, value: RichText) -> None:
-        if value.key != "text":
-            raise ValueError("RichText's key is must be `text`")
-        self.__text = value
-        self["bulleted_list_item"].update(self.__text)
-
-    @property
-    def children(self) -> Children:
-        return self["bulleted_list_item"]["children"]
-
-    @children.setter
-    def children(self, value: Children) -> None:
-        self["bulleted_list_item"].update(value)
-
-
-class BulletedList(Children):
-    """
-    BulletedList
-    BulletedList property values of block
-
-    Attributes
-    ----------
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    def __init__(
-        self,
-        *item: BulletedListItem,
-    ):
-        """
-        Parameters
-        ----------
-        *item : BulletedListItem
-            items of bulleted list item
-        """
-        super().__init__(*item)
-
-
-class NumberedListItem(Block):
-    """
-    NumberedListItem
-    NumberedListItem property values of block
-
-    Attributes
-    ----------
-    *text : Text or RichText
-        text
-    children : Children, optional
-        children
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "type": "numbered_list_item",
-        "numbered_list_item": {
-            "text": [],
-        },
-    }
-
-    def __init__(
-        self,
-        *text: Union[Text, RichText],
-        children: Optional[Children] = None,
-    ):
-        """
-        Parameters
-        ----------
-        *text : Text or RichText
-            text
-        children : Children, optional
-            children
-        """
-        super().__init__()
-        base = []  # Aggregate Texts
-        for t in text:
-            if isinstance(t, RichText):
-                base.extend(list(t[t.key]))
-            elif isinstance(t, Text):
-                base.append(t)
-            else:
-                raise ValueError(
-                    f"Expected type is `RichText` or `Text`, but {type(t)} is given"
-                )
-        self.__text = RichText(key="text", *base)
-        self["numbered_list_item"].update(self.__text)  # Add Texts with RichText Style
-        if children is not None:
-            self["numbered_list_item"].update(
-                children
-            )  # if children exists, Add Chilren
-
-    @property
-    def text(self) -> RichText:
-        return self.__text
-
-    @text.setter
-    def text(self, value: RichText) -> None:
-        if value.key != "text":
-            raise ValueError("RichText's key is must be `text`")
-        self.__text = value
-        self["numbered_list_item"].update(self.__text)
-
-    @property
-    def children(self) -> Children:
-        return self["numbered_list_item"]["children"]
-
-    @children.setter
-    def children(self, value: Children) -> None:
-        self["numbered_list_item"].update(value)
-
-
-class NumberedList(Children):
-    """
-    NumberedList
-    NumberedList property values of block
-
-    Attributes
-    ----------
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    def __init__(
-        self,
-        *item: NumberedListItem,
-    ):
-        """
-        Parameters
-        ----------
-        *item : NumberedListItem
-            items of numbered list item
-        """
-        super().__init__(*item)
-
-
-class ToDo(Block):
-    """
-    ToDo
-    ToDo property values of block
-
-    Attributes
-    ----------
-    text : RichText
-        Text in the to_do block
-    checked : bool, default=False
-        Whether the to_do is checked or not
-    children : Children
-        Any nested children blocks of the to_do block
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict, bool]] = {
-        "type": "to_do",
-        "to_do": {
-            "text": [],
-            "checked": False,
-        },
-    }
-
-    def __init__(
-        self,
-        *text: Union[Text, RichText],
-        checked: bool = False,
-        children: Optional[Children] = None,
-    ):
-        """
-        Parameters
-        ----------
-        *text : Text or RichText
-            Text in the to_do block
-        checked : bool, default=False
-            Whether the to_do is checked or not
-        children : Children, optional
-            Any nested children blocks of the to_do block
-        """
-        super().__init__()
-        base = []  # Aggregate Texts
-        for t in text:
-            if isinstance(t, RichText):
-                base.extend(list(t[t.key]))
-            elif isinstance(t, Text):
-                base.append(t)
-            else:
-                raise ValueError(
-                    f"Expected type is `RichText` or `Text`, but {type(t)} is given"
-                )
-        self.__text = RichText(key="text", *base)
-        self["to_do"].update(self.__text)  # Add Texts with RichText Style
-        self["to_do"]["checked"] = checked  # Add Checked
-        if children is not None:
-            self["to_do"].update(children)  # if children exists, Add Chilren
-
-    @property
-    def text(self) -> RichText:
-        return self.__text
-
-    @text.setter
-    def text(self, value: RichText) -> None:
-        if value.key != "text":
-            raise ValueError("RichText's key is must be `text`")
-        self.__text = value
-        self["to_do"].update(self.__text)
-
-    @property
-    def children(self) -> Children:
-        return self["to_do"]["children"]
-
-    @children.setter
-    def children(self, value: Children) -> None:
-        self["to_do"].update(value)
-
-    @property
-    def checked(self) -> bool:
-        return self["to_do"]["checked"]
-
-    @checked.setter
-    def checked(self, value: bool) -> None:
-        self["to_do"]["checked"] = value
-
-    @checked.deleter
-    def checked(self) -> None:
-        self["to_do"]["checked"] = False
-
-
-class ToDoList(Children):
-    """
-    ToDoList
-    ToDoList property values of block
-
-    Attributes
-    ----------
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    def __init__(
-        self,
-        *item: ToDo,
-    ):
-        """
-        Parameters
-        ----------
-        *item : ToDo
-            items of todo item
-        """
-        super().__init__(*item)
-
-
-class Toggle(Block):
-    """
-    Toggle
-    Toggle property values of block
-
-    Attributes
-    ----------
-    text : RichText
-        Rich text in the toggle block
-    children : Children
-        Any nested children blocks of the toggle block
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict, bool]] = {
-        "type": "toggle",
-        "toggle": {
-            "text": [],
-            "children": [],
-        },
-    }
-
-    def __init__(
-        self,
-        *text: Union[Text, RichText],
-        children: Optional[Children] = None,
-    ):
-        """
-        Parameters
-        ----------
-        *text : RichText
-            Rich text in the toggle block
-        children : Children
-            Any nested children blocks of the toggle block
-        """
-        super().__init__()
-        base = []  # Aggregate Texts
-        for t in text:
-            if isinstance(t, RichText):
-                base.extend(list(t[t.key]))
-            elif isinstance(t, Text):
-                base.append(t)
-            else:
-                raise ValueError(
-                    f"Expected type is `RichText` or `Text`, but {type(t)} is given"
-                )
-        self.__text = RichText(key="text", *base)
-        self["toggle"].update(self.__text)  # Add Texts with RichText Style
-        if children is not None:
-            self["toggle"].update(children)  # if children exists, Add Chilren
-
-    @property
-    def text(self) -> RichText:
-        return self.__text
-
-    @text.setter
-    def text(self, value: RichText) -> None:
-        if value.key != "text":
-            raise ValueError("RichText's key is must be `text`")
-        self.__text = value
-        self["toggle"].update(self.__text)
-
-    @property
-    def children(self) -> Children:
-        return self["toggle"]["children"]
-
-    @children.setter
-    def children(self, value: Children) -> None:
-        self["toggle"].update(value)
-
-
-class Code(Block):
-    """
-     Code
-     Code property values of block
-
-    Attributes
-    ----------
-    text : RichText
-        Rich text in code block
-    language : str, optional
-        Coding language in code block
-    valid_language : Literal
-        Possible values for language
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "type": "code",
-        "code": {
-            "text": [],
-            "language": "",
-        },
-    }
-
-    def __init__(
-        self,
-        *text: Union[Text, RichText],
-        language: Optional[LANGUAGES] = None,
-    ):
-        """
-        Parameters
-        ----------
-        *text : Text or RichText
-            Rich text in code block
-        language : str, optional
-            Coding language in code block
-        """
-        super().__init__()
-        base = []  # Aggregate Texts
-        for t in text:
-            if isinstance(t, RichText):
-                base.extend(list(t[t.key]))
-            elif isinstance(t, Text):
-                base.append(t)
-            else:
-                raise ValueError(
-                    f"Expected type is `RichText` or `Text`, but {type(t)} is given"
-                )
-        self.__text = RichText(key="text", *base)
-        self["code"].update(self.__text)  # Add Texts with RichText Style
-        if language is not None:
-            self["code"]["language"] = language  # Add Language
-
-    @property
-    def valid_language(self):
-        return LANGUAGES
-
-    @property
-    def text(self) -> RichText:
-        return self.__text
-
-    @text.setter
-    def text(self, value: RichText) -> None:
-        if value.key != "text":
-            raise ValueError("RichText's key is must be `text`")
-        self.__text = value
-        self["code"].update(self.__text)
-
-    @property
-    def language(self) -> bool:
-        return self["code"]["language"]
-
-    @language.setter
-    def language(self, value: LANGUAGES) -> None:
-        self["code"]["language"] = value
-
-    @language.deleter
-    def language(self) -> None:
-        self["code"]["language"] = ""
-
-
-class Embed(Block):
-    """
-     Embed
-     Embed property values of block
-
-    Attributes
-    ----------
-    url : str
-        Link to website the embed block will display
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "type": "embed",
-        "embed": {
-            "url": "",
-        },
-    }
-
-    def __init__(
-        self,
-        url: str,
-    ):
-        """
-        Parameters
-        ----------
-        url : str
-            Link to website the embed block will display
-        """
-        super().__init__()
-        self["embed"]["url"] = url
-
-    @property
-    def url(self) -> str:
-        return self["embed"]["url"]
-
-    @url.setter
-    def url(self, value: str) -> None:
-        self["embed"]["url"] = value
-
-    @url.deleter
-    def url(self) -> None:
-        self["embed"]["url"] = ""
-
-
-class Image(Block):
-    """
-     Image
-     Image property values of block
-     Includes supported image urls (i.e. ending in .png, .jpg, .jpeg, .gif, .tif, .tiff, .bmp, .svg, or .heic)
-
-    Attributes
-    ----------
-    url : str
-        Link to website the image block will display
-    type_ : 'external' or 'file', default='external'
-        Type of this file object. Possible values are: 'external', 'file'
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "type": "image",
-        "image": {
-            "type": "external",
-            "external": {
-                "url": "",
-            },
-        },
-    }
-
-    def __init__(
-        self,
-        *caption: Union[Text, RichText],
-        url: Optional[str] = None,
-        type_: Literal["external", "file"] = "external",
-        file: Optional[FileObject] = None,
-    ):
-        """
-        Parameters
-        ----------
-        *caption : Text or RichText
-            Caption of the image block
-        url : str, optional
-            Link to website the image block will display
-        type_ : 'external' or 'file', default='external'
-            Type of this file object. Possible values are: 'external', 'file'
-        file : FileObject, optional
-            FileObject
-        """
-        super().__init__()
-        base = []  # Aggregate Texts
-        for t in caption:
-            if isinstance(t, RichText):
-                base.extend(list(t[t.key]))
-            elif isinstance(t, Text):
-                base.append(t)
-            else:
-                raise ValueError(
-                    f"Expected type is `RichText` or `Text`, but {type(t)} is given"
-                )
-        self.__caption = RichText(key="caption", *base)
-
-        if url is None and file is None:
-            raise ValueError("Either url or file should be not None")
-        elif file is not None:
-            self.__file = file
-        elif url is not None:
-            ext = url.split(".")[-1]
-            if ext not in IMAGE_EXT:
-                raise ValueError(
-                    """Includes supported image urls,
-                    (i.e. ending in .png, .jpg, .jpeg, .gif, .tif, .tiff, .bmp, .svg, or .heic)"""
-                )
-            self.__file = FileObject(type_=type_, url=url)
-        self["image"] = self.__file
-        self["image"].update(self.__caption)
-
-    @property
-    def caption(self) -> RichText:
-        return self.__caption
-
-    @caption.setter
-    def caption(self, value: RichText) -> None:
-        self.__caption = value
-        self["image"].update(self.__caption)
-
-    @caption.deleter
-    def caption(self) -> None:
-        self.__caption = RichText()
-        self["image"].update(self.__caption)
-
-    @property
-    def type_(self) -> str:
-        return self.__file.type_
-
-    @type_.setter
-    def type_(self, value: Literal["external", "file"]) -> None:
-        self.__file.type_ = value
-        self["image"] = self.__file
-
-    @type_.deleter
-    def type_(self) -> None:
-        self.__file.type_ = "external"
-        self["image"] = self.__file
-
-    @property
-    def url(self) -> str:
-        return self.__file.url
-
-    @url.setter
-    def url(self, value: str) -> None:
-        self.__file.url = value
-
-    @url.deleter
-    def url(self) -> None:
-        self.__file.url = ""
-
-
-class Video(Block):
-    """
-     Video
-     Video property values of block
-
-    Attributes
-    ----------
-    *caption : Text or RichText
-            Caption of the video block
-    url : str
-        Link to website the video block will display
-    type_ : 'external' or 'file', default='external'
-        Type of this file object. Possible values are: 'external', 'file'
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "type": "video",
-        "video": {
-            "type": "external",
-            "external": {
-                "url": "",
-            },
-        },
-    }
-
-    def __init__(
-        self,
-        *caption: Union[Text, RichText],
-        url: Optional[str] = None,
-        type_: Literal["external", "file"] = "external",
-        file: Optional[FileObject] = None,
-    ):
-        """
-        Parameters
-        ----------
-        url : str, optional
-            Link to website the video block will display
-        type_ : 'external' or 'file', default='external'
-            Type of this file object. Possible values are: 'external', 'file'
-        file : FileObject, optional
-            FileObject
-        """
-        super().__init__()
-        base = []  # Aggregate Texts
-        for t in caption:
-            if isinstance(t, RichText):
-                base.extend(list(t[t.key]))
-            elif isinstance(t, Text):
-                base.append(t)
-            else:
-                raise ValueError(
-                    f"Expected type is `RichText` or `Text`, but {type(t)} is given"
-                )
-        self.__caption = RichText(key="caption", *base)
-
-        if url is None and file is None:
-            raise ValueError("Either url or file should be not None")
-        elif file is not None:
-            self.__file = file
-        elif url is not None:
-            self.__file = FileObject(type_=type_, url=url)
-        self["video"] = self.__file
-        self["video"].update(self.__caption)
-
-    @property
-    def caption(self) -> RichText:
-        return self.__caption
-
-    @caption.setter
-    def caption(self, value: RichText) -> None:
-        self.__caption = value
-        self["video"].update(self.__caption)
-
-    @caption.deleter
-    def caption(self) -> None:
-        self.__caption = RichText()
-        self["video"].update(self.__caption)
-
-    @property
-    def type_(self) -> str:
-        return self.__file.type_
-
-    @type_.setter
-    def type_(self, value: Literal["external", "file"]) -> None:
-        self.__file.type_ = value
-        self["video"] = self.__file
-
-    @type_.deleter
-    def type_(self) -> None:
-        self.__file.type_ = "external"
-        self["video"] = self.__file
-
-    @property
-    def url(self) -> str:
-        return self.__file.url
-
-    @url.setter
-    def url(self, value: str) -> None:
-        self.__file.url = value
-
-    @url.deleter
-    def url(self) -> None:
-        self.__file.url = ""
-
-
-class File(Block):
-    """
-     File
-     File property values of block
-
-    Attributes
-    ----------
-    url : str
-        Link to website the video block will display
-    type_ : 'external' or 'file', default='external'
-        Type of this file object. Possible values are: 'external', 'file'
-    caption : RichText
-        Caption of the file block
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "type": "file",
-        "file": {
-            "type": "external",
-            "external": {
-                "url": "",
-            },
-        },
-    }
-
-    def __init__(
-        self,
-        *caption: Union[Text, RichText],
-        url: Optional[str] = None,
-        type_: Literal["external", "file"] = "external",
-        file: Optional[FileObject] = None,
-    ):
-        """
-        Parameters
-        ----------
-        *caption : Text or RichText
-            Caption of the file block
-        url : str, optional
-            Link to website the video block will display
-        type_ : 'external' or 'file', default='external'
-            Type of this file object. Possible values are: 'external', 'file'
-        file : FileObject, optional
-            FileObject
-        """
-        super().__init__()
-        base = []  # Aggregate Texts
-        for t in caption:
-            if isinstance(t, RichText):
-                base.extend(list(t[t.key]))
-            elif isinstance(t, Text):
-                base.append(t)
-            else:
-                raise ValueError(
-                    f"Expected type is `RichText` or `Text`, but {type(t)} is given"
-                )
-        self.__caption = RichText(key="caption", *base)
-
-        if url is None and file is None:
-            raise ValueError("Either url or file should be not None")
-        elif file is not None:
-            self.__file = file
-        elif url is not None:
-            self.__file = FileObject(type_=type_, url=url)
-        self["file"] = self.__file
-        self["file"].update(self.__caption)
-
-    @property
-    def caption(self) -> RichText:
-        return self.__caption
-
-    @caption.setter
-    def caption(self, value: RichText) -> None:
-        self.__caption = value
-        self["file"].update(self.__caption)
-
-    @caption.deleter
-    def caption(self) -> None:
-        self.__caption = RichText()
-        self["file"].update(self.__caption)
-
-    @property
-    def type_(self) -> str:
-        return self.__file.type_
-
-    @type_.setter
-    def type_(self, value: Literal["external", "file"]) -> None:
-        self.__file.type_ = value
-        self["file"] = self.__file
-
-    @type_.deleter
-    def type_(self) -> None:
-        self.__file.type_ = "external"
-        self["file"] = self.__file
-
-    @property
-    def url(self) -> str:
-        return self.__file.url
-
-    @url.setter
-    def url(self, value: str) -> None:
-        self.__file.url = value
-
-    @url.deleter
-    def url(self) -> None:
-        self.__file.url = ""
-
-
-class Pdf(Block):
-    """
-     Pdf
-     Pdf property values of block
-
-    Attributes
-    ----------
-    url : str
-        Link to website the pdf block will display
-    type_ : 'external' or 'file', default='external'
-        Type of this file object. Possible values are: 'external', 'file'
-    caption : RichText
-        Caption of the pdf block
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "type": "pdf",
-        "pdf": {
-            "type": "external",
-            "external": {
-                "url": "",
-            },
-        },
-    }
-
-    def __init__(
-        self,
-        *caption: Union[Text, RichText],
-        url: Optional[str] = None,
-        type_: Literal["external", "file"] = "external",
-        file: Optional[FileObject] = None,
-    ):
-        """
-        Parameters
-        ----------
-        *caption : Text or RichText
-            Caption of the pdf block
-        url : str, optional
-            Link to website the pdf block will display
-        type_ : 'external' or 'file', default='external'
-            Type of this file object. Possible values are: 'external', 'file'
-        file : FileObject, optional
-            FileObject
-        """
-        super().__init__()
-        base = []  # Aggregate Texts
-        for t in caption:
-            if isinstance(t, RichText):
-                base.extend(list(t[t.key]))
-            elif isinstance(t, Text):
-                base.append(t)
-            else:
-                raise ValueError(
-                    f"Expected type is `RichText` or `Text`, but {type(t)} is given"
-                )
-        self.__caption = RichText(key="caption", *base)
-
-        if url is None and file is None:
-            raise ValueError("Either url or file should be not None")
-        elif file is not None:
-            self.__file = file
-        elif url is not None:
-            ext = url.split(".")[-1]
-            if ext not in ("pdf",):
-                raise ValueError("Includes supported image urls, (i.e. ending in .pdf)")
-            self.__file = FileObject(type_=type_, url=url)
-        self["pdf"] = self.__file
-        self["pdf"].update(self.__caption)
-
-    @property
-    def caption(self) -> RichText:
-        return self.__caption
-
-    @caption.setter
-    def caption(self, value: RichText) -> None:
-        self.__caption = value
-        self["pdf"].update(self.__caption)
-
-    @caption.deleter
-    def caption(self) -> None:
-        self.__caption = RichText()
-        self["pdf"].update(self.__caption)
-
-    @property
-    def type_(self) -> str:
-        return self.__file.type_
-
-    @type_.setter
-    def type_(self, value: Literal["external", "file"]) -> None:
-        self.__file.type_ = value
-        self["pdf"] = self.__file
-
-    @type_.deleter
-    def type_(self) -> None:
-        self.__file.type_ = "external"
-        self["pdf"] = self.__file
-
-    @property
-    def url(self) -> str:
-        return self.__file.url
-
-    @url.setter
-    def url(self, value: str) -> None:
-        self.__file.url = value
-
-    @url.deleter
-    def url(self) -> None:
-        self.__file.url = ""
-
-
-class Bookmark(Block):
-    """
-     Bookmark
-     Bookmark property values of block
-
-    Attributes
-    ----------
-    url : str
-        Link to website the Bookmark block will display
-    caption : RichText
-        Caption of the pdf block
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "type": "bookmark",
-        "bookmark": {
-            "url": "",
-        },
-    }
-
-    def __init__(
-        self,
-        url: str,
-        *caption: Union[Text, RichText],
-    ):
-        """
-        Parameters
-        ----------
-        url : str
-            Link to website the bookmark block will display
-        *caption : Text or RichText
-            Caption of the bookmark block
-        """
-        super().__init__()
-        base = []  # Aggregate Texts
-        for t in caption:
-            if isinstance(t, RichText):
-                base.extend(list(t[t.key]))
-            elif isinstance(t, Text):
-                base.append(t)
-            else:
-                raise ValueError(
-                    f"Expected type is `RichText` or `Text`, but {type(t)} is given"
-                )
-        self.__caption = RichText(key="caption", *base)
-        self["bookmark"]["url"] = url
-        self["bookmark"].update(self.__caption)
-
-    @property
-    def caption(self) -> RichText:
-        return self.__caption
-
-    @caption.setter
-    def caption(self, value: RichText) -> None:
-        self.__caption = value
-        self["bookmark"].update(self.__caption)
-
-    @caption.deleter
-    def caption(self) -> None:
-        self.__caption = RichText()
-        self["bookmark"].update(self.__caption)
-
-    @property
-    def url(self) -> str:
-        return self["bookmark"]["url"]
-
-    @url.setter
-    def url(self, value: str) -> None:
-        self["bookmark"]["url"] = value
-
-    @url.deleter
-    def url(self) -> None:
-        self["bookmark"]["url"] = ""
-
-
-class Equation(Block):
-    """
-    Equation
-    Equation property values of block
-
-    Attributes
-    ----------
-    expression : str
-        A KaTeX compatible string
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "type": "equation",
-        "equation": {"expression": ""},
-    }
-
-    def __init__(
-        self,
-        expression: str,
-    ):
-        """
-        Parameters
-        ----------
-        expression : str
-            A KaTeX compatible string
-        """
-        super().__init__()
-        self["equation"]["expression"] = expression
-
-    @property
-    def expression(self) -> str:
-        return self["equation"]["expression"]
-
-    @expression.setter
-    def expression(self, value: str) -> None:
-        self["equation"]["expression"] = value
-
-    @expression.deleter
-    def expression(self) -> None:
-        self["equation"]["expression"] = ""
-
-
-class Divider(Block):
-    """
-    Divider
-    Divider property values of block
-
-    Attributes
-    ----------
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "type": "divider",
-        "divider": {},
-    }
-
-    def __init__(self):
-        super().__init__()
-
-
-class TableOfContents(Block):
-    """
-    TableOfContents
-    TableOfContents property values of block
-
-    Attributes
-    ----------
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "type": "table_of_contents",
-        "table_of_contents": {},
-    }
-
-    def __init__(self):
-        super().__init__()
-
-
-class BreadCrumb(Block):
-    """
-    BreadCrumb
-    BreadCrumb property values of block
-
-    Attributes
-    ----------
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "type": "breadcrumb",
-        "breadcrumb": {},
-    }
-
-    def __init__(self):
-        super().__init__()
-
-
-class Column(Block):
-    """
-    Column
-    Column property values of block
-    This must have at least one child
-
-    Attributes
-    ----------
-    children : Children
-        blocks in a single column
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "type": "column",
-        "column": {},
-    }
-
-    def __init__(
-        self,
-        *block: Block,
-    ):
-        """
-        Parameters
-        ----------
-        *block : Block
-            blocks in a single column.
-        """
-        super().__init__()
-        if len(block) < 1:
-            raise ValueError("This must have at least one block")
-        self.__children = Children(*block)
-        self["column"] = self.__children
-
-    @property
-    def children(self) -> Children:
-        return self.__children
-
-    @children.setter
-    def children(self, value: Children) -> None:
-        self.__children = value
-        self["column"] = self.__children
-
-
-class ColumnList(Block):
-    """
-    Column
-    Column property values of block
-    This must have at least 2 columns
-
-    Attributes
-    ----------
-    children : Children
-        List of columns in the column_list block
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "type": "column_list",
-        "column_list": {},
-    }
-
-    def __init__(
-        self,
-        *column: Column,
-    ):
-        """
-        Parameters
-        ----------
-        *column : Column
-            columns in the column_list block.
-        """
-        super().__init__()
-        if len(column) < 2:
-            raise ValueError("This must have at least 2 columns")
-        self.__children = Children(*column)
-        self["column_list"] = self.__children
-
-    @property
-    def children(self) -> Children:
-        return self.__children
-
-    @children.setter
-    def children(self, value: Children) -> None:
-        self.__children = value
-        self["column_list"] = self.__children
-
-
-class LinkToPage(Block):
-    """
-     LinkToPage
-     LinkToPage property values of block
-
-    Attributes
-    ----------
-    type_ : str
-        Type of this link to page object. Possible values are: "page", "database"
-    id_ : str
-        Identifier for a page or a database page
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "type": "link_to_page",
-        "link_to_page": {},
-    }
-
-    def __init__(
-        self,
-        type_: Literal["page", "database"],
-        id_: str,
-    ):
-        """
-        Parameters
-        ----------
-        type_ : 'page' or 'database'
-            Type of this link to page object. Possible values are: "page", "database"
-        id_ : str
-            Identifier for a page or a database page, URL style is ok.
-        """
-        super().__init__()
-        if type_ not in ("page", "database"):
-            raise ValueError("type_ must be `page` or `database`")
-        id_ = parse_id(id_=id_, type_=type_)
-        type_ = f"{type_}_id"
-        self["link_to_page"]["type"] = type_
-        self["link_to_page"][type_] = id_
-
-    @property
-    def type_(self) -> str:
-        return self["link_to_page"]["type"]
-
-    @property
-    def id_(self) -> str:
-        return self["link_to_page"][self.type_]
-
-
-class OriginalSynced(Block):
-    """
-     OriginalSynced
-     OriginalSynced property values of block
-     To create a synced_block, the developer needs to create an original synced block.
-     Developers will be able to identify the original synced_block because it does not "sync_from" any other block
-     (synced_from property is set to null)
-
-    Attributes
-    ----------
-    children : Children
-        Any nested children blocks of the synced_block block.
-        These blocks will be synced across this block and references to this synced_block
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "type": "synced_block",
-        "synced_block": {
-            "synced_from": None,
-        },
-    }
-
-    def __init__(
-        self,
-        *child: Block,
-    ):
-        """
-        Parameters
-        ----------
-        *child : Block
-            Any nested children blocks of the synced_block block.
-            These blocks will be synced across this block and references to this synced_block
-        """
-        super().__init__()
-        self.__children = Children(*child)
-        self["synced_block"].update(self.__children)
-
-    @property
-    def children(self) -> Children:
-        return self.__children
-
-    @children.setter
-    def children(self, children: Children) -> None:
-        self.__children = children
-        self["synced_block"].update(self.__children)
-
-
-class ReferenceSynced(Block):
-    """
-    ReferenceSynced
-    ReferenceSynced property values of block
-    To sync the content of the original synced_block with another synced_block,
-    the developer simply needs to refer to that synced_block using the synced_from property
-
-    Attributes
-    ----------
-    block_id : str
-        Identifier of an original synced_block
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "type": "synced_block",
-        "synced_block": {
-            "synced_from": {
-                "block_id": "",
-            },
-        },
-    }
-
-    def __init__(
-        self,
-        block_id: str,
-    ):
-        """
-        Parameters
-        ----------
-        block_id : str
-            Identifier of an original synced_block
-        """
-        super().__init__()
-        self["synced_block"]["synced_from"]["block_id"] = block_id
-
-    @property
-    def block_id(self) -> str:
-        return self["synced_block"]["synced_from"]["block_id"]
-
-    @block_id.setter
-    def block_id(self, block_id: str) -> None:
-        self["synced_block"]["synced_from"]["block_id"] = block_id
-
-    @block_id.deleter
-    def block_id(self) -> None:
-        self["synced_block"]["synced_from"]["block_id"] = ""
-
-
-class TableRow(Block):
-    """
-    TableRow
-    TableRow property values of block
-
-    Attributes
-    ----------
-    cells : list of list of RichText
-        Array of cell contents in horizontal display order.
-        Each cell itself is an array of rich text objects
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "type": "table_row",
-        "table_row": {
-            "cells": [],
-        },
-    }
-
-    def __init__(
-        self,
-        *cell: RichText,
-    ):
-        """
-        Parameters
-        ----------
-        *cell : RichText
-            Array of cell contents in horizontal display order.
-            Each cell itself is an array of rich text objects
-        """
-        super().__init__()
-        for c in cell:
-            self["table_row"]["cells"].append(c[c.key])
-
-    def __add__(self, other: Union[RichText, List[RichText]]):
-        if isinstance(other, list):
-            self.extend(other)
-            return self
-        self.append(other)
-        return self
-
-    def __iadd__(self, other: Union[RichText, List[RichText]]):
-        return self.__add__(other)
-
-    def __len__(self):
-        return len(self["table_row"]["cells"])
-
-    @property
-    def cells(self) -> List[List[Dict]]:
-        return self["table_row"]["cells"]
-
-    @cells.deleter
-    def cells(self) -> None:
-        while len(self["table_row"]["cells"]) > 0:
-            _ = self["table_row"]["cells"].pop()
-
-    def append(self, cell: RichText) -> None:
-        """
-        append(cell: RichText)
-            Append RichText to existing cells
-
-        Parameters
-        ----------
-        cell : RichText
-            RichText you append to cells
-        """
-        self["table_row"]["cells"].append(cell[cell.key])
-
-    def extend(self, cells: List[RichText]) -> None:
-        """
-        extens(cells: List[RichText])
-            Append RichText to existing cells
-
-        Parameters
-        ----------
-        cells : list of RichText
-            List of RichText you append to cells
-        """
-        for cell in cells:
-            self["table_row"]["cells"].append(cell[cell.key])
-
-    def insert(self, index: int, cell: RichText) -> None:
-        """
-        insert(index: int, cell: RichText)
-            Append RichText to existing cells
-
-        Parameters
-        ----------
-        index : int
-            Index you insert Text into
-        cell : RichText
-            RichText you insert into cells
-        """
-        self["table_row"]["cells"].insert(index, cell[cell.key])
-
-    def pop(self, index=None):
-        """
-        pop(index: int)
-            Pop cell from existing cells
-
-        Parameters
-        ----------
-        index : int, default=None
-            Index of cell you pop
-        """
-        self["table_row"]["cells"].pop(index)
-
-
-class Table(Block):
-    """
-    Table
-    Table property values of block
-    Tables are parent blocks for table row children. They can only contain children of type table_row.
-    When creating a table block via the Append block children endpoint,
-    the table must have at least 1 table_row whose cells array has the same length as the table_width
-
-    Attributes
-    ----------
-    table_width : int
-        Number of columns in the table.
-        Note that this cannot be changed via the public API once a table is created
-    children : Children
-        table row children
-    has_column_header : bool, default=False
-        Whether or not the table has a column header.
-        If true, the first row in the table will appear visually distinct from the other rows.
-    has_row_header : bool, default=False
-        Whether or not the table has a header row.
-        If true, the first column in the table will appear visually distinct from the other columns.
-
-    Methods
-    -------
-    clear()
-        Clear data of title
-    json()
-        Return this class as dictionary
-    """
-
-    TEMPLATE: Dict[str, Union[str, Dict]] = {
-        "type": "table",
-        "table": {
-            "table_width": 1,
-            "has_column_header": False,
-            "has_row_header": False,
-        },
-    }
-
-    def __init__(
-        self,
-        table_width: int,
-        *table_row: TableRow,
-        has_column_header: bool = False,
-        has_row_header: bool = False,
-    ):
-        """
-        Parameters
-        ----------
-        table_width : int
-            Number of columns in the table.
-            Note that this cannot be changed via the public API once a table is created
-        *table_row : TableRow
-            table row children
-        has_column_header : bool, default=False
-            Whether or not the table has a column header.
-            If true, the first row in the table will appear visually distinct from the other rows.
-        has_row_header : bool, default=False
-            Whether or not the table has a header row.
-            If true, the first column in the table will appear visually distinct from the other columns.
-        """
-        super().__init__()
-        for table_row_ in table_row:  # Validate length of table row
-            if len(table_row_) != table_width:
-                raise ValueError(
-                    "table_width must be equal to table_row size, "
-                    f"expected {table_width} but {len(table_row_)} is given"
-                )
-        self["table"]["table_width"] = table_width
-        self["table"]["has_column_header"] = has_column_header
-        self["table"]["has_row_header"] = has_row_header
-        self.__children = Children(*table_row)
-        self["table"].update(self.__children)
-
-    @property
-    def table_width(self) -> int:
-        return self["table"]["table_width"]
-
-    @property
-    def has_column_header(self) -> bool:
-        return self["table"]["has_column_header"]
-
-    @has_column_header.setter
-    def has_column_header(self, value: bool) -> None:
-        self["table"]["has_column_header"] = value
-
-    @property
-    def has_row_header(self) -> bool:
-        return self["table"]["has_row_header"]
-
-    @has_row_header.setter
-    def has_row_header(self, value: bool) -> None:
-        self["table"]["has_column_header"] = value
-
-    @property
-    def children(self) -> Children:
-        return self.__children
-
-    @children.setter
-    def children(self, value: Children) -> None:
-        self.__children = value
-        self["table"].update(self.__children)
-
-    @children.deleter
-    def children(self) -> None:
-        self.__children = Children()
-        self["table"].update(self.__children)
