+++ notion_extensions/base/props/block/block.py
@@ -1,0 +1,14 @@
+from ..common import BaseProps
+
+__all__ = [
+    "Block",
+]
+
+
+class Block(BaseProps):
+    """
+    Block Object
+    """
+
+    def __init__(self):
+        super().__init__()
