+++ notion_extensions/base/props/block/paragraph.py
@@ -1,0 +1,123 @@
+from typing import Dict, List, Union
+
+from .block import Block
+from ..common import Text, RichText
+
+__all__ = [
+    "Paragraph",
+]
+
+
+class Paragraph(Block):
+    """
+    Paragraph
+    Paragraph property values of block
+
+    Attributes
+    ----------
+
+    Methods
+    -------
+    clear()
+        Clear data of title
+    json()
+        Return this class as dictionary
+    """
+
+    TEMPLATE: Dict[str, Union[str, Dict]] = {
+        "object": "block",
+        "type": "paragraph",
+        "paragraph": {
+            "text": [],
+        },
+    }
+
+    def __init__(
+        self,
+        *text: Union[Text, RichText],
+    ):
+        super().__init__()
+        base = []
+        for t in text:
+            if isinstance(t, RichText):
+                base.extend(list(t[t.key]))
+            elif isinstance(t, Text):
+                base.append(t)
+            else:
+                raise ValueError(
+                    f"Expected type is `RichText` or `Text`, but {type(t)} is given"
+                )
+        self.__texts = RichText(key="text", *base)
+
+        self.update(
+            {
+                "object": "block",
+                "type": "paragraph",
+                "paragraph": self.__texts,
+            },
+        )
+
+    def __add__(self, other: Union[Text, List[Text]]):
+        if isinstance(other, list):
+            self.extend(other)
+            return self
+        self.append(other)
+        return self
+
+    def __iadd__(self, other: Union[Text, List[Text]]):
+        return self.__add__(other)
+
+    def append(self, text: Text) -> None:
+        """
+        append(text: Text)
+            Append Text to existing list of Text
+
+        Parameters
+        ----------
+        text : Text
+            Text you append to RichText
+        """
+        self.__texts.append(text)
+        self["paragraph"] = self.__texts
+
+    def extend(self, texts: List[Text]) -> None:
+        """
+        extens(texts: Text)
+            Append Text to existing list of Text
+
+        Parameters
+        ----------
+        text : list of Text
+            List of text you append to RichText
+        """
+        self.__texts.extend(texts)
+        self["paragraph"] = self.__texts
+
+    def insert(self, index: int, text: Text) -> None:
+        """
+        insert(index: int, text: Text)
+            Append Text to existing list of Text
+
+        Parameters
+        ----------
+        index : int
+            Index you insert Text into
+        text : Text
+            Text you insert into RichText
+        """
+        self.__texts.insert(index, text)
+        self["paragraph"] = self.__texts
+
+    def pop(self, index=None):
+        """
+        pop(text: Text)
+            Pop Text to existing list of Text
+
+        Parameters
+        ----------
+        index : int, default=None
+            Text you pop from RichText
+        """
+        item = self.__texts.pop(index)
+        self["paragraph"] = self.__texts
+        return item
